Started making a GUI
Ok, before you throw up all over the floor because of the colors, just know, I only left them there so you could understand the spacing of the webpage. I added a random div with a height of like 10,000 pixels so you can see how when you scroll, the text input stays where it is, and the fixed content on the right, which will eventually be for like different chatrooms, export options, options, etc. So the only thing that scrolls is the chat. Hope it good!