Local storage works & cleaned up theme-changer
I made the buttons for the themes dropdown auto-generate depending on the themeChanger.themes. This way we don't have to manually add the buttons. Also, the local storage was not working for a specific reason: the client.py is set up right now to open the GUI on a random port number. While that's temporary, that makes each different client you open on your pc a different webpage technically. So in client.py, I just added a commented line that has a hardcoded port number if you want to test the savedOptions in action. I also cleaned up your implementation of the local storage to have a savedOptions class so we can save more than just the themes in a clean way.